backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
passlib[bcrypt]
pydantic
bcrypt
python-jose[cryptography]
cloudinary
python-multipart


